+++ entregas/garciaDavid/reto002/Reto002Poo/.idea/.gitignore
@@ -1,0 +1,3 @@
+# Default ignored files
+/shelf/
+/workspace.xml
