+++ entregas/garciaDavid/reto002/Reto002Poo/src/imprimircursor.java
@@ -1,0 +1,6 @@
+public class imprimircursor {
+    static void imprimeCursor(int cur) {
+        String[] cursor = {" - "};
+        System.out.print(cursor[cur]);
+    }
+}
