+++ entregas/garciaDavid/reto002/Reto002Poo/src/limpiarTablero.java
@@ -1,0 +1,6 @@
+public class limpiarTablero extends limpiar{
+    static void limpiarPantalla() {
+
+        System.out.print("\033[H\033[2J");
+    }
+}
